+++ Makefile
@@ -89,6 +89,7 @@ DEFINES =
 INCLUDES += $(shell sh -c 'xml2-config --cflags')
 INCLUDES += $(shell sh -c 'pkg-config --cflags glib-2.0')
 INCLUDES += -Ilib/stringencoders
+INCLUDES += -Itools/include
 
 EXTRA_LIBS += $(shell sh -c 'xml2-config --libs')
 EXTRA_LIBS += $(shell sh -c 'pkg-config --libs glib-2.0')$
@@ -115,16 +116,20 @@ endif
 
 market_EXTRA_DEPS += lib/die.o
 market_EXTRA_DEPS += tools/sim/engine.o
+market_EXTRA_DEPS += tools/fix/fix_common.o
 market_EXTRA_LIBS += -lm
 
 trader_EXTRA_DEPS += lib/die.o
+trader_EXTRA_DEPS += tools/fix/fix_common.o
 
 fix_client_EXTRA_DEPS += lib/die.o
 fix_client_EXTRA_DEPS += tools/fix/test.o
+fix_client_EXTRA_DEPS += tools/fix/fix_common.o
 fix_client_EXTRA_LIBS += -lm
 
 fix_server_EXTRA_DEPS += lib/die.o
 fix_server_EXTRA_DEPS += tools/fix/test.o
+fix_server_EXTRA_DEPS += tools/fix/fix_common.o
 
 fast_client_EXTRA_DEPS += lib/die.o
 fast_client_EXTRA_DEPS += tools/fast/test.o
@@ -137,6 +142,7 @@ fast_parser_EXTRA_DEPS += lib/die.o
 fast_parser_EXTRA_DEPS += tools/fast/test.o
 
 forts_EXTRA_DEPS += lib/die.o
+forts_EXTRA_DEPS += tools/fix/fix_common.o
 
 tape_EXTRA_DEPS += tools/tape/builtin-check.o
